feat: add Projects page
Introduce the Projects page with a title, card components, and styling. Include a script to enable horizontal scrolling in the chip interactions section. Utilize Svelte fade transitions for in and out effects with a delay of 500ms. Incorporate external link icons for GitHub and other links. Adjust styles for better responsiveness.